Extracted find_or_create_build_dir func
